airflow/requirements.txt
alembic==1.4.2
apache-airflow==1.10.12
apispec==1.3.3
argcomplete==1.11.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
cached-property==1.5.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.32
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
docutils==0.16
email-validator==1.1.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14
gunicorn==19.10.0
idna==2.9
importlib-metadata==1.6.1
iso8601==0.1.12
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.3
joblib==0.15.1
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.5.0
lockfile==0.12.2
lxml==4.5.1
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==7.0.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.5
pandas==0.25.3
pendulum==1.4.4
Pillow==7.1.2
prison==0.1.3
psutil==5.7.0
psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
pyrsistent==0.16.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
requests-file==1.5.1
setproctitle==1.1.10
six==1.15.0
soupsieve==2.0.1
psycopg2==2.8.6
SQLAlchemy==1.3.17
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
cattrs==1.0
text-unidecode==1.3
thrift==0.13.0
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.9
Werkzeug==0.16.1
wincertstore==0.2
WTForms==2.3.1
zipp==3.1.0
zope.deprecation==4.4.0
